Fixup issue with reverting `mode-line-format'
There's an issue with reverting `mode-line-format', which seems to
affect at least dired buffers.  Easiest way to replicate would be to:

1. Open up a dired buffer in ~

2. From there jump to any directory

3. Enable feebleline-mode

3. Jump back to the dired buffer located at ~

4. Disable feebleline-mode

This would result in `mode-line-format' not being reverted back to the
value of `feebleline--mode-line-format-previous' inside the dired buffer
opened in step 2., instead `mode-line-format' would still be nil, and so
there would be no mode-line in that buffer.  After these steps it is
very easy to break the mode-lines of every buffer by just toggling on
and off feebleline-mode in different buffers.

Seems traversing only the visible windows using `walk-windows' isn't
enough in some cases, like on dired buffers, to revert
`mode-line-format'.  Instead traversing all buffers returned by
`buffer-list' seems do the trick.

Ignore ephemeral and uninteresting buffers

Traversing all buffers causes an issue with a mode-line appearing in buffers
that originally did not have a mode-line assigned to them.  By ignoring
uninteresting buffers (buffers whose name starts with a space), this issue
should be solved.